Add distributed Covid tracing app (#15)
This is a very rough first take of a Hydroflow app that crosses network
boundaries. Lots of stuff is bad here: the abstractions are wonky, the encoding
allocates too much, the message-passing is fragile, but I wanted to get this
merged so that stuff could be improved incrementally as we figure out what
actually good abstractions are here.