refactor(hydro_lang): directly manipulate IR in `KeyedStream` (#2149)
Because the ordering guarantees differ between `Stream` and
`KeyedStream`, we will want to emit different IR metadata depending on
which collection is used. This makes our current approach to
`KeyedStream` wrapping over an unordered `Stream` a bit clumsy, because
then we don't get to control the IR metadata.

Now, because `KeyedStream` directly emits the IR, we will be able to
customize its metadata to capture inherent partial orders. For example,
in the old code, calling `.batch` would potentially shuffle ordered
values with the same key since the underlying stream was unordered. Now,
batch will have special behavior for keyed streams with ordered values
to preserve the guaranteed partial order.